feat: allow disabling savestates
Some libretro cores do not support savestates, and including them in the UI causes users to falsely believe they will work as the screenshot is taken but the file is not written. Rather than just disable the screenshot functionality, remove the options altogether to avoid confusion.

To remove the load/save options from the minarch menu, create a disable_savestates file in the emulator's pak folder.