feat: Add support for Doom via the PRBOOM libretro core
Doom games can be added to the `/Roms/Doom (PRBOOM)` folder. While wads can be loaded directly, using an m3u folder structure is recommended:

```text
/Roms/Doom (PRBOOM)/Mod Name/
    Mod Name.m3u
    game1.wad
    game2.wad
    doom.version
```

The `m3u` file should contain the wad names in the following format:

```shell
game1.wad
game2.wad
```

The `prboom.wad` file should be present in the `/Bios/PRBOOM` folder. You may additionally add any of the official doom IWADs to the `/Bios/PRBOOM` folder, as well as any mp3 music you wish. The following official IWAD names are supported in the specified order by the PRBOOM core:

- `doom2f.wad`
- `doom2.wad`
- `plutonia.wad`
- `tnt.wad`
- `freedoom2.wad`
- `doom.wad`
- `doomu.wad`
- `freedoom1.wad`
- `freedoom.wad`
- `doom1.wad`

As the WADs are loaded in the above fixed order, it may be necessary to add a `doom.version` file to the folder to force-load a specific IWAD. This file should contain a single line with the version name, for example:

```text
doom1
```

A corresponding `doom1` folder should be present in the `/Bios/PRBOOM` folder, and this can contain all the assets that should be loaded for that version of the game. For instance, the following might be in the `/Bios/PRBOOM/doom1` folder:

```shell
doom1.wad
prboom.wad
intro.mp3
e1m1.mp3
```

If the `prboom.wad` file is not present in the `/Bios/PRBOOM/doom1` folder, it will be copied from the `/Bios/PRBOOM` folder.

PRBOOM does not support savestates.