fix: crash when available pill width is negative
caveat: The pill size is calculated on the theoretical thumbnail width (e.g. 100% of screen width), but the actual game art might be smaller (due to the top and bottom menu elements taking up space, and the specific aspect ratio of your game art). As a result, 100% width will always hide the pill, but not really fullscreen the thumbnail as you might expect.