feat: provide the ability to hide entry names when browsing the root directory
Some themes may wish to monopolize the screen, displaying the current entry name in the background image directly instead of via a selected text. This change provides the ability to show/hide all entry names when browsing the root directory - when top->path == SDCARD_PATH - via a setting that is on by default. This has no effect on any subdirectories or the Tools directory.